Support 64 bit values in unsigned_int object

+++ include/fc/interprocess/container.hpp
@@ -113,7 +113,7 @@ namespace fc {
        inline void pack( Stream& s, const bip::vector<T,A...>& value, uint32_t _max_depth=FC_PACK_MAX_DEPTH ) {
           FC_ASSERT( _max_depth > 0 );
           --_max_depth;
-          pack( s, unsigned_int((uint32_t)value.size()), _max_depth );
+          pack( s, unsigned_int(value.size()), _max_depth );
           auto itr = value.begin();
           auto end = value.end();
           while( itr != end ) {
